feat!: remove LookupKubeconfigPath
